Merge survey results page.
Survey results had 2 pages, one with the summary and one with the
individual responses. I've merged them together and patched them to work
with us using event_id instead of survey_id now in SurveyInput.

Still to be done: Patch survey submission so that it uses event_id
instead of survey_id.