Individual survey result add name, fix no response.
The individual survey result page is the page where you view the survey
responses of a student student. For some reason, the student is not
identified on the page, so I've added the student's name as a header.

The "no response" marker wasn't showing up due to a missing condition,
this was fixed and the marker was changed in order to be more obvious
that it's not user input.