fixed some bugs in the result view
= there are differences in the list of returned group members between
mixeval and rubrics. Mixeval has a list of users that includes roles
that are indexed with 0 which means it knows that user can have
multiple roles. However Rubrics doesn't. The reason was the list
was generated with sql join.
	- I have patched the problem for rubrics in the
getEventGroupMembersNoTutors function in the groups_members model.
In the future we should add the user model to the groups_members
so that the problem can be permanently fixed.

= I have also fixed the rubrics summary table to have each criterion
to be out of the correct mark. Previously it was made out of the
criterion number.

= I have also removed colours from the criteria and question
numbering for mixeval and rubrics. The reason is some of the
colours used were hard to see with the yellow background.