Allow instructors to add courses.
UI changes so that instructors can be assigned a faculty. This will
allow instructors to select a department for courses.

And a small enhancement where if an instructor is adding a course,
they're automatically pre-selected as an instructor.

Also updated ACL so that it's easier to diff. Was not sure if the
instructors course add permissions were properly updated.