Changed blinking note and Submitted evals
Changed the NOTE blinking text which appeared in Simple Eval Form to ensure that it does not blink anymore.

Submitted Evals appear under the upcoming section if the current date is between the release dates. This is
to ensure that the student can edit a submitted evaluation(still needs to be implemented completely).