Allow students to edit submitted evals.
Changed the "submitted" section in student home so that if result
release hasn't happened yet and we're still within the release period,
then a link to edit the eval is provided.

Note that this partially reverts the home_controller in the previous
commit which put editable evals into the "due" section.