Add local vocabulary class for translation
Some time the displayed word may not fit the application. Added a local
vocabulary that allow user to set the word to what they want. e.g. at
UBC, the administration is not deep down to department level. So
changing department text to faculty and creating one to one
faculty-department is easier for administration.