Survey View: Now shows survey questions/responses
The Add Event page links to Survey View for previews.  This turned out
not to be very useful since the view only just showed the survey's
title, creator, and create/modified dates. I've taken some of the code
from the question editing section and added in a question display to the
view section.

This should allow a better rewriting of the question editing section now
that it doesn't have to worry about toggling editable or not on and off.