fixed a user form bug & add loop to assign faculty to super admin
= there was a bug introduced in an earlier commit that did not
allow super admins to be added if no faculties are selected.
However the faculty field is greyed out. It has now been fixed.
Adding super admins no longer required the faculty field to be
not empty.

= when a new faculty is added it is assigned to all super admins
in the system. However there was a bug where it only assigns the
last super admin on the list to the new faculty. This is fine
when there is only one super admin in the system, but when there
are more, the other super admins will not have access to the new
faculty. This has now been fixed by fixing how the array is
created.

= I have also added a similar loop in the users controller add
function. When a super admin is added to the system, it will go
through all the faculties in the system and create a relationship
between the faculties and the new super admin.