made minor cosmetic changes
= sorted the list of other recipients for writing email in
alphabetical order so that it is user to find the desired users.
In production the list would be very long because the list
includes every user in the system except for the users already
added to the email.

= changed the field label for faculty for users/edit to
instructor/admin only from admin only because it has been changed
that instructors now belong to faculties.

= for simple_eval_form added a space between the number of
empty comment fields and the phrase "comments to be filled"