Fix TeamMaker execution on Linux & error display.
TeamMaker couldn't be executed because it didn't have the execute flag,
chmod +x fixed it.

TeamMaker was just failing silently without informing the user. As a
quick fix, this commit checks the TeamMaker return status, and if it
isn't successful, it'll at least tell the user that TeamMaker execution
was at fault.