Fix point dist on editing a simple evaluation.
When a student edit one of their previously submitted simple evaluation,
the point distribution calculated by javascript ("Points
Allocated/Total") doesn't get initialized properly. The 'Submit
Evaluation' button also isn't available.

Don't know why the javascript on this page is all done in php echos and
don't know why it's at the top of the page instead of at the button.
These should rectified later, but for now, to limit the scope of my
changes, I've just moved the initial javascript update call to the
bottom of the file.