fixed bug where instructors can view private evaluation templates
The bug was created when we started to place instructors into
faculties. Previously we assumed only the admins belong to
faculties. Due to this assumption instructors are given access
to all private templates created by other instructors in their
faculty. This has now been fixed. Instructors are only given
access to private templates that they created or created by an
instructor they are sharing a course with.