filtered out add buttons for invalid courses
For invalid courses, I have filtered out the add users, events, etc
links from the indices. I have also changed a line in the "How to use
Lists in iPeer" section of the indices to say Left-Clicking the
light yellow instead of gray since the colour of the background has
been changed.