UserCourse test failing due to missing fixtures.
Apparently, even an empty test case needs to have all its fixtures
declared.

This is another of the silently failing test cases that Jenkins doesn't
catch. The only indication that we have a failure is when the # of test
cases run drops significantly for no reason. Probably because the test
case result parser rejects the result file as invalid.