fixed some bugs in the user controller
If the user check off the send email notification checkbox, an
invalid bug will appear. It expects a field named full_name, but
no fields by that name exist anymore. I have fixed it by changing
it to take the concatenation of the first name and last name.

The Reset Password does not work because it was going through the
beforeValidate function but it is a different structure than the
add & edit action which was what the beforeValidate function was
expecting. To fix it I turned off validation for the save function,
since no validation is required.

I have also made a minor change to the survey group sets index.
Originally the Released? field was one of the search options. For
it does not make sense to search 'Yes' or 'No' in that field. I
was going to make it a dropdown menu, but the map functionality for
ajax expects a char(1) attribute but availability is an int(1)
attribute. Instead I have made it a custom field so that it does
not appear as one of the search options.