modified enrolling and unenrolling students
= from the class list the delete link no longer deletes the user
from the system. Instead it unenrols the student from the course.

= an option is given for importing users. Instructors and admins
can now choose to merge their new class list with the class list
they have previously imported. Another option is to update the
class list and remove all students that were on the list in the
previous import but not on the new list.