refactored groups import view
It now looks very similar to users import. Another benefit is the
code is a lot cleaner, now that we have gotten rid of the tables.
Another reason for refactoring was previously the list of courses
is generated with the course selection box element which includes
an option called all courses which breaks if it is selected. Now
the list of courses is generated with cakephp magic and only includes
the list of courses the user has access to.