fixed user validation
The structure for user data is different among add & edit users,
resetting passwords, and adding super admin (install). For the fix
I accomodate for the differences in the array structure and not
validate at all when resetting passwords. Another bug I have fixed
is for users edit. The bug was when a user was editted without
any courses selected, an invalid variable error appears. I just moved
the fix from users add to users edit.