removed email confirmation for evaluation submission
There is a bug when the sendConfirmationEmail function is called,
which is when a user with an email submits an evaluation or survey.
The bug was the undefined property Email. I have removed it because
it is a little redundant to send an email confirmation when you can
visually see your evaluation and survey submission by clicking on the
links under the Submitted sections.