fixed bugs from unenrolling students for mixeval results
I have changed the colour for "Entered" to darker green so that it
can be read easier. I have also shorten the save buttons and moved
the note to below the buttons similar to rubrics. The biggest change
I made is the filtering of the submissions for the mixeval results.
The results are only returned to the view if they were not created
by an unenrolled student.