fixed bugs caused by unenrolling students for rubrics
I've added a filter to filter out all submissions made by students
that have been unenrolled from the course already. All sums and
averages also only include submissions from students still in the
course.

I've also fixed a bug for the home index when a student is unenrolled.
There are cases when "8 of 7 students" completed the evaluations get
printed. I've fixed it by changing the sql query in the event model.

I've also changed the colour for the rubrics evaluation "Saved"
to green from red to keep the colour coding consistent. In mixed
evaluations it is also green. I've also fixed the probrlem where
the statement that comments are required is printed even if the
event does not require comments.