fixed some bugs & outdated instructions
I've fixed the bug where instructors can view emails in the
users/view even though they are not suppose to like how it was
taken out in go_to_class_list.

Updated the Short Help on the home index. The biggest change is
the instructions to view the User Documentation on trac instead
of the iPeer Tutorial Wizard.

Changed the width of the button so that all the words can appear
on it for forms like groups/add. I have also fixed the checkbox
label css problem on the users/add and users/edit.

lastly, I have removed the instructions to come up with usernames
with a minimum length of 6 characters which have been taken out
of the actual form.