fixed unenrolling students via users/edit
Previously, when a student is unenrolled via users/edit by
unchecking the course, they are only unenrolled from the classlist
but not taken out of the groups and surveygroups of the course.
This has been done by doing a check before anything new is saved.
During the process I have moved the getEnrolledCourses function
from UserEnrol model to User model. This also lead me to move the
tests for it from UserEnrol to Users test cases.

I have changed the link in goToClassList from Delete to Drop to
avoid the confusion of whether the link deletes the user from the
system or drops the student from the course.