refactored mixeval results page
I have some of the data processing from the model to the controller
so that queries will be done in their respective models instead
of all in the evaluation component. The view has also been changed
to display evaluators and/or evaluatees that are not part of the
group in blue to let the instructor know the member no longer
belongs to the group.