fixed bugs in mixeval results
A bug was introduced where names were not printed in the summary
table but instead the word "Array" was printed. This was due to the
addition of an array added to the data to include class="blue" when
needed (eg. when the student is no longer in the group). This has
now been fixed and unused variables/codes have also been removed.