combined evaltools index with showAll
Previously the index for evaltools have the option of passing
a parameter such as rubrics to be redirected to rubrics/index.
However this "functionality" is no longer used so I have
consolidated the showAll function with the index function and
remove the option of passing in an argument for index.

During the process I have also simplified the code by using
Cakephp's findAllBy functions instead of using find with conditions.
I have also decrease the number of lines of code to check whether
any events using a particular template has any submissions.

Furthermore, I have included a return statement after a redirect
for the mixevals controller.