Add mixeval_question_types table.
This should make it easier to add question types to mixeval_questions.
There are a few feature requests on trac about new question types, this
is preliminary work meant to help implement these new question types.

Note that the question_type column in mixeval_questions is now
mixeval_question_type_id, it's a foreign key mapped to the the new
mixeval_question_types table.

Also added some instructions to the sample mixeval questions for testing
that instruction display is working in mixeval.