Fix #460
I've moved the url link for the course homepage from the course
index to the course home page. Instructors can now have access
to the course home page from there. The Web column for course
index has now been removed.

We have found that if a url without http is filled in, the system
will append localhost... to the beginning of the url if the user
is using a a brower on a Mac. To fix this I added a custom
validation function for homepage which checks to make sure
the url starts with http. If not, it will append "http://" to
it as a prefix before saving. I have also added an extra validation
for the homepage to make sure it is in a valid URL format.