Fix #477
Instructors can now add events with titles that have already been
used whether the duplicated title appears in the course or not.
However, if the title exists in another event in the course, a
warning will appear (similar to the username error) in the users
add form.

I have also fixed the users enrol for duplicated usernames in
the users add form to return an error if the student or tutor
is already in the course.