Ticket #476
If the instructor did not select a department when creating a course,
the system will not put the course under ALL the departments that
belongs to ALL the faculties that ALL the instructors in the course
are in. This change will allow the faculty admin to have access to
the course even if a department is not selected. However if a
department is selected, the course will only be placed under the
selected departments.