changed the view for users and groups import
I have changed the help-text for the checkboxes to be more precise
and have added the update or "append" option to the groups import
form (however it is yet to be implemented).

I have also fixed the css for the checkbox to align with its label.
Lastly, I have added comments in the groups controller regarding
the changes I would make in a future commit to include the update
or "append" option and to make group# optional.