Fix misaligned send email notification label.
The send email notification checkbox's label was floating to the left of
the checkbox like it's supposed to. This was due to it having a custom
div class of "email" so it didn't get the general rules for all labels.
Fixed by removing the custom div class and some extraneous css.