Added the Auto-Released Results option for events (simple & rubric)
The database will have to be reinstalled to include the new field
"auto_release" for the events table. With this change I have also
changed the events index to now show comments required or self
evaluation, instead it is showing the state for evaluation results,
including not yet, manual, closed, n/a (survey), or auto. We will
still need to include the auto release messages to the mixeval
results.

Side Note: I have also shortened the text for "string" for the
ajaxlist to be less than or equal to 160 characters.