initial commit for moving student from one course to another
At this point we can successfully move a student's survey
submissions from Source course to Destination course if
a survey event using the same template is already created in
the destination course. More work has to be done for mass import
and to include functionalities such as enroling the student
in the destination course as well.