included form validation and data check to courses move
Since the form data does not get saved directly to the database we
cannot use the cakephp validation. I have instead checked that all
five fields have a selection (all required). If even one is missing,
the user will be redirected back to courses/move with a message
saying that all fields are required.

I have also included a check to see if a submission has been made to
the destination survey already. If a submission has been made the user
will also get an error message saying that the student seleted has
already submitted the survey.

Minor change: instead of using find in the move function, I have
changed them to use functions that are from their respective
models.