finished implementing the basics for moving students' surveys
The "Move Student" button is only available to "move" the selected
student when a select fields are not empty.

I have also included an option for the user to choose whether they
want to move or copy the student. To move, it means that the survey
results would be "cut and paste" to the destination course's event
and the student would be unenrolled from the source course. If the
copy option is selected a copy of the submission and inputs would
be made with the destination's event id. For both choices, the
student would be enrolled to the destination course if not already.