modified courses/move user interface
Changed some of the field labels to be singular so that it is more
clear what each field is for. I have also included a couple
instructions to make the feature more user friendly.

For the code that populates the select field, I have moved the find
queries to their respective models. To do this I had to either
create new functions or edit existing functions.