New mixeval add now fully functional.
Supports likert questions now and can actually save the mixeval properly
to database.

Saving to the database is a bit complicated due to the need to use
transactions. Since CakePHP does not support nested transactions and
since saveAll creates a transaction by default, saveAll cannot be used
in my transactional save. Each Mixeval, MixevalQuestion, and
MixevalQuestionDesc has to be saved individually with a regular model
save call.

It would've been nice to have the likert marks distributed properly but
since that doesn't impact functionality, it was pushed back later.

Note that likert questions right now is more of a generic template for
other questions that might need to save descriptors too, e.g.: multiple
choice. Haven't figured out how to nicely abstract the shared qualities
of these questions while having the ability to customize.

It would be nice to be able to decare questions as "Master" like in
surveys, allowing you to load them into an evaluation. But that is for
later too.