changed the form field for adding instructors to courses
Ticket #444. The field has been changed to a dropdown menu with a
link the user will click to add the selected instructor. When the
link is clicked a textbox with the instructor would be added and
the instructor would be removed from the dropdown menu. The
instructor's textbox will disappear if 'X' is selected and they
would be readded to the dropdown menu.

Due to this change. The Instructor array in the $this->data has
changed slightly, therefore I reformat the data in the add and edit
functions. I have also added code to delete removed instructors for
courses edit.

With the change of the format for Instructor, I had to change the
data for the courses edit test in the courses controller or else
some of the tests would break.