Mixeval edit now works for question deletion.
Question deletion is done by comparing the data currently in the
database and what the form submitted. If the form submitted data is
missing IDs, then the question or question desc that the ID refers to is
deleted. This is still part of the save transaction, so it will be
rolled back if an error was encountered.