finished implementing users/merge
I have added the link to the page to the top of users index with
a merge icon. Furthermore I have modified the jQuery for the page
to show the process of searching for users so that users know that
the server is not just hanging if no results are returned.