Fix v1_controller test can't find host in console
This only happens when we run the test cases on the command line. The
failing test in question give a url "v1/users...". The key here is that
when running the tests on the web interface, CakePHP's router properly
handles that as a relative URL and returns something like
"http://localhost/v1/users/...". Unfortunately, when this test is run
from the command line, we don't have the benefits of Router's relative
URL awareness, and have to manually construct the URL ourselves, so we
get a wrong url in the form "http://localhostv1/users/...".

I've added a check that should prevent this from happening in the
future.