Fix delta_5 query failing in newer MySQL version.
Newer MySQL versions in strict mode now does validation on DATETIME
fields to make sure that they're valid dates. This means that the
default value used and inserted by iPeer2, '0000-00-00', will fail
validation when we try to upgrade the database.

It's too late to fix this for delta_5, as it's already been released and
we should touch it as little as possible, but at least we can fix the
invalid dates in the 3.1 upgrade script. Luckily, we removed a lot of the
SysParameter entries and won't have to fix as many of them.