implemented a check for the mixeval form
If a student has submitted and want to resubmit the evaluation, we
need a check to make sure all required questions are answered. If
not we may have instances where students clear their orignal answers
for required questions and resubmit. This will result in incomplete
submissions.

I have implemented a jQuery that is only done if a submission has
already been done. If the check finds that a required question is
unanswered, the evaluation will not be resubmitted and will ask
the student to answer all required questions before submitting.