reimplemented enable_details for simple and rubric eval
The feature which was already implemented for simple and rubric
evaluations results earlier have been slightly modified. This
allows instructors to have the option to show detailed results
to the students or a simplified version where only averages are
given.

One change that was made is the field is now named enable_details
instead of results, so that it is more clear what the is the field's
use. I've also set the default option in the events add and edit
page to be detailed results.