modified mixeval form's evaluation details
Due to the change in the structure of the mixeval form, which
includes not using the standardtable class. By having a yellow
theme for the detaisl was a little out of place. To change that
I have modified the details

At the very top, we now have the evaluation's name hyphen the
group's name. Then we have some Notes which includes the due
date, explanation for the red asterix, penalty policy, event
description (if one exists), and a message telling the student that
all required questions must be answered to resubmit the evaluation.
If no submissions have been made, the student does not need to
answer all required questions before saving.