modified mixeval view page
Instead of having duplicated code, I have modified mixeval view
to use the element mixevals/view_mixeval_details that has already
been modified to use the new structure for the mixeval form.
The biggest change I made was to assign a user['id']=0 whenever
a user id is used in the form.