Allow custom delta files in db patcher.
The upgrade_base's patchDb function only takes delta_*.sql files, where
* is a number. I needed it to run additional delta files according to
iPeer version and so added the $additionalDeltas param, where you can
state additional delta files for the patcher to run after the main delta
files has finished.

Note that I removed the $dbconfig param in patchDb cause it didn't seem
to be used anyway.

Also added exceptions to upgrade_300 so that it doesn't run when we're
already in an iPeer v3.0.x database. In fact, had to add this exception
to the upgrade_controller too so that the correct database versions will
be displayed when upgrading to v3.1.

Had to force a SysParameter reload on login as I was having problems
where the SysParameter changed in the database but wasn't reflected in
the cache even after I've logged out.