refactored students' evaluation results view for mix evaluations
Since the "evaluation form" for mix evaluation has been modified
to not use the accordion and the mix evaluation has been restructured
that Likert, Sentence, and Paragraph questions can be in any order.

This means that the evaluation results view that shows Section 1
and Section 2 no longer applies. The evaluation results view for
students have been restructured to be the list of questions
with the evaluators' answers listed below them with bullet points.

The list of questions and answeres is generated in a elements
evaluations/mixeval_details, which can also be used for the
instructors' evaluation results.