fixed submitting mixeval evaluation bugs
Group events were being marked at "to review" even when not every
one in the group has submitted. The bug was the number of group
members being used to compared was the number of members in the
curent user's evaluation which is definitely less than the number
of group members especially when tutors or no self evaluation is
involved. This has been fixed by changing the number of group
members to number of users in the group in the groups_members
table.

Another bug found was when a user submits their mix evaluation,
a lot of times they are redirected back to the evaluation with
no messages so it seems like the page just refreshed. But really
the answers were saved and submitted. The reason was a successful
submission message only appeared if the user is the last one to
submit in the group (aka the group event can be marked as "to
review"). I have fixed it by setting a successful message whether
the group event is marked as "to review" or not.

We will have to take a look at whether a check will have to be done
when a user is unenrolled AND taken out of a group. If a student is
taken out, it may result in the whole group have submitted and the
evaluation should be set to "to review".