implemented condensed version of student's view of mixeval results
Like the detailed version it has the evaluation event detail and
rating tables at the top. One of the difference is the condensed
version only show averages of likert questions (both required and
unrequired). I think showing the comments (written ques) defeats
the purpose of avoiding the concern mentioned in Ticket #407.