modified mix evaluation results for instructors
Like the student's mix evaluation results view, it no longer make
sense to use Section 1 and Section 2 to separate the text and Likert
questions because the actual evaluation form doesn't use sections
anymore. The evaluation results now uses the element mixval_details.
The only difference between the instructor's and student's view is
that the evaluator's name is printed for the instructor while the
answers are shown anonymously.