made a small addition to the mix eval results for instructors
Included the addition of the colour blue to evaluator's name that
no longer belongs to the group. This feature was in the previous
mix evaluation results, but was left out when the new format
was implemented.

I've also removed a few functions that were used by mix evaluation
results but no longer used including getPoints, renderQuestionResult,
getMixevalResultTable, and getResultTableHeader from the
evaluation helper.