made some minor changes throughout the application
If a student submits after an instructor releases the comments or
answers from the individual group results, the button shows
unrelease comments or grades instead of releasing. I've fixed it
by setting the buttons to release when a minimum of one submission
is not released.

Included a javascript confirm for the drop action in go_to_class_list
so that students would not be accidentally unenrolled.

I've changed the wording for the confirm message for users/message
to have a gentler tone.

Added * to required questions for mix evaluation results.