fixed release buttons for simple and rubric evaluation results
Previously, the release buttons' actions (eg. release/unrelease)
depends on a boolean that is determined by the last
evaluation_simple or evaluation_rubric entry in the data array.
I've fixed that by multiplying all the booleans for the individual
release booleans for each user.

This is to accomodate for the possibility of a student submitting
after the instructor has already released the grades and/or
comments. If that's the case the button will be to release instead
of possibly unrelease.

I have also removed the formatMixevalEvaluationResultsMatrix
function and its test since it's no longer used.