Added reminder for evaluation submission
1. Fade in when students click on Save - Modified
rubric_eval_form.ctp to show an alert when students click
on save button. The animation tells them to click on the submit
button.

2. Pop up when students leave the page without clicking submit -
Shows a popup when students try to exit the page or click on
another link without submitting the evaluation

3. Email reminder before deadline - Modified events_controller.php
to allow the instructor to select the frequency of reminder emails
(Disable,1...7 days). Modified send_emails.php to filter the to
list based on reminder date and evaluation submission.

2.