set mixeval questions to default to required.
This change was made for the convenience of the instructor. This is
so that they don't have to check off each question as required.

I have also added help-text for likert questions reminding the user
that unrequired likert questions would not be counted toward the
total ratings in the results.