made some minor modifications to groups/import
In the Instructions section, the statement that tells instructors
to only use usernames or student numbers but not both has been
modified to say to use a student identifier which can be student
numbers or usernames.

For the formatting section, the Username column has been changed
to Student Identifier to avoid any unnecessary confusion.

I have added a help-text for the student identifier form field to
remind the user to select the student identifier used in the CSV
file.

The default student identifier has been changed to username from
student number.

For the error message when an unknown user is placed in the CSV
file, I have added an additional question which asks whether the
user has chosen the correct student identifier.