fixed mixeval questions numbering
If the initial mixeval only has one question, and is being editted
the numbering of the questions restart at number 1. This bug has
been fixed. The reason was javascript translated index 0 to null.

On the side, I have also fixed mixeval index's total mark column.
When there are no questions in the template, the total mark column
is left blank, so I have made it 0 if no questions are found.