Changed send_emails in order to consider edit events
Change send_emails.php so that the function
reminderFilter() gets the latest list of groups
in the event rather than the one that was saved in the past.
This is useful since a group might be taken out of an event
once the event has been made. The group that gets taken out should
not be sent any message.